+++ railties/lib/rails/info_controller.rb
@@ -19,12 +19,12 @@ class Rails::InfoController < Rails::ApplicationController # :nodoc:
   end
 
   def routes
-    if path = params[:path]
-      path = URI::DEFAULT_PARSER.escape path
-      normalized_path = with_leading_slash path
+    if query = params[:query]
+      query = URI::DEFAULT_PARSER.escape query
+
       render json: {
-        exact: match_route { |it| it.match normalized_path },
-        fuzzy: match_route { |it| it.spec.to_s.match path }
+        exact: matching_routes(query: query, exact_match: true),
+        fuzzy: matching_routes(query: query, exact_match: false)
       }
     else
       @routes_inspector = ActionDispatch::Routing::RoutesInspector.new(_routes.routes)
@@ -33,11 +33,31 @@ class Rails::InfoController < Rails::ApplicationController # :nodoc:
   end
 
   private
-    def match_route
-      _routes.routes.filter_map { |route| route.path.spec.to_s if yield route.path }
-    end
+    def matching_routes(query:, exact_match:)
+      return [] if query.blank?
 
-    def with_leading_slash(path)
-      ("/" + path).squeeze("/")
+      normalized_path = ("/" + query).squeeze("/")
+      query_without_url_or_path_suffix = query.gsub(/(\w)(_path$)/, '\1').gsub(/(\w)(_url$)/, '\1')
+
+      _routes.routes.filter_map do |route|
+        route_wrapper = ActionDispatch::Routing::RouteWrapper.new(route)
+
+        if exact_match
+          match = route.path.match(normalized_path)
+          match ||= (query_without_url_or_path_suffix === route_wrapper.name)
+        else
+          match = route_wrapper.path.match(query)
+          match ||= route_wrapper.name.include?(query_without_url_or_path_suffix)
+        end
+
+        match ||= (query === route_wrapper.verb)
+
+        unless match
+          controller_action = URI::DEFAULT_PARSER.escape(route_wrapper.reqs)
+          match = exact_match ? (query === controller_action) : controller_action.include?(query)
+        end
+
+        route_wrapper.path if match
+      end
     end
 end